+++ lib/pronto/gitlab.rb
@@ -2,6 +2,7 @@ module Pronto
   class Gitlab
     def initialize(repo)
       @repo = repo
+      @config = Config.new
       @comment_cache = {}
     end
 
@@ -22,8 +23,9 @@ module Pronto
     private
 
     def slug
+      return @config.gitlab_slug if @config.gitlab_slug
       @slug ||= begin
-        host = URI.split(endpoint)[2, 2].compact.join(':')
+        host = URI.split(@config.gitlab_api_endpoint)[2, 2].compact.join(':')
         slug = @repo.remote_urls.map do |url|
           match = /.*#{host}(:|\/)(?<slug>.*).git/.match(url)
           match[:slug] if match
@@ -33,15 +35,8 @@ module Pronto
     end
 
     def client
-      @client ||= ::Gitlab.client(endpoint: endpoint, private_token: private_token)
-    end
-
-    def private_token
-      ENV['GITLAB_API_PRIVATE_TOKEN']
-    end
-
-    def endpoint
-      ENV['GITLAB_API_ENDPOINT']
+      @client ||= ::Gitlab.client(endpoint: @config.gitlab_api_endpoint,
+                                  private_token: @config.gitlab_api_private_token)
     end
 
     Comment = Struct.new(:sha, :note, :path, :line) do